+++ lib/src/component_animated_builder.dart
@@ -22,7 +22,7 @@ class AnimatedMvvmContent<P extends Presenter, M extends MVVMModel>
 
   @override
   State<MVVMContent> createState() =>
-      _MVVMSingleTickerProviderContentState<Presenter, MVVMModel>();
+      _MVVMSingleTickerProviderContentState<P, M>(animListener);
 }
 
 class _MVVMSingleTickerProviderContentState<P extends Presenter,
@@ -31,6 +31,9 @@ class _MVVMSingleTickerProviderContentState<P extends Presenter,
     implements MVVMView {
   AnimationController _controller;
   bool hasInit = false;
+  final MvvmAnimationListener<P, M> animListener;
+
+  _MVVMSingleTickerProviderContentState(this.animListener);
 
   @override
   void didChangeDependencies() {
@@ -62,69 +65,21 @@ class _MVVMSingleTickerProviderContentState<P extends Presenter,
 
   P get presenter => PresenterInherited.of<P, M>(context).presenter;
 
-  MvvmContext get mvvmContext => MvvmContext(context);
+  MvvmContext get mvvmContext =>
+      MvvmContext(context, animationController: _controller);
 
   MvvmContentBuilder<P, M> get builder =>
       PresenterInherited.of<P, M>(context).builder;
 
   @override
-  Widget build(BuildContext context) => builder(
-        MvvmContext(context, animationController: _controller),
-        presenter,
-        presenter.viewModel,
-      );
+  Widget build(BuildContext context) =>
+      builder(mvvmContext, presenter, presenter.viewModel);
 
   @override
   Future<void> refreshAnimation() async {
-    widget.animListener(
-      MvvmContext(context, animationController: _controller),
-      presenter,
-      presenter.viewModel,
-    );
+    animListener(mvvmContext, presenter, presenter.viewModel);
   }
 }
-
-//class AnimatedMvvmContent<P extends Presenter, M extends MVVMModel> extends MVVMContent {
-//
-//  final MvvmAnimationControllerBuilder singleAnimController;
-//  final MvvmAnimationListener<P, M> animListener;
-//
-//  AnimatedMvvmContent({
-//    Key key,
-//    @required MvvmContentBuilder<P, M> builder,
-//    @required this.singleAnimController,
-//    this.animListener,
-//  }) : super(key: key, builder: builder);
-//
-//  @override
-//  _MVVMSingleTickerProviderContentState<P, M> createState() =>
-//    _MVVMSingleTickerProviderContentState<P, M>(_builder, this.singleAnimController, this.animListener);
-//}
-//
-//
-//class _MVVMSingleTickerProviderContentState<P extends Presenter, M extends MVVMModel> extends _MVVMContentState with SingleTickerProviderStateMixin {
-//
-//  final MvvmAnimationControllerBuilder animationControllerBuilder;
-//  final MvvmAnimationListener<P, M> animListener;
-//  AnimationController _controller;
-//
-//  _MVVMSingleTickerProviderContentState(dynamic builder, this.animationControllerBuilder, this.animListener) : super(builder);
-//
-//  @override
-//  void didChangeDependencies() {
-//    super.didChangeDependencies();
-//    _controller = animationControllerBuilder(this);
-//  }
-//
-//  @override
-//  Widget build(BuildContext context) =>
-//    _builder(MvvmContext(context, animationController: _controller), _presenter, _presenter.viewModel);
-//
-//  @override
-//  refreshAnimation() async {
-//    animListener(MvvmContext(context, animationController: _controller), _presenter, _presenter.viewModel);
-//  }
-//}
 
 /// -----------------------------------------------
 /// MULTIPLE ANIMATIONS CONTENT WIDGET
@@ -142,7 +97,7 @@ class MultipleAnimatedMvvmContent<P extends Presenter, M extends MVVMModel>
 
   @override
   _MVVMMultipleTickerProviderContentState<P, M> createState() =>
-      _MVVMMultipleTickerProviderContentState<P, M>();
+      _MVVMMultipleTickerProviderContentState<P, M>(animListener);
 }
 
 class _MVVMMultipleTickerProviderContentState<P extends Presenter,
@@ -151,6 +106,9 @@ class _MVVMMultipleTickerProviderContentState<P extends Presenter,
     implements MVVMView {
   List<AnimationController> _controller;
   bool hasInit = false;
+  final MvvmAnimationListener<P, M> animListener;
+
+  _MVVMMultipleTickerProviderContentState(this.animListener);
 
   @override
   void didChangeDependencies() {
@@ -188,60 +146,11 @@ class _MVVMMultipleTickerProviderContentState<P extends Presenter,
       PresenterInherited.of<P, M>(context).builder;
 
   @override
-  Widget build(BuildContext context) => builder(
-        MvvmContext(context, animationsControllers: _controller),
-        presenter,
-        presenter.viewModel,
-      );
+  Widget build(BuildContext context) =>
+      builder(mvvmContext, presenter, presenter.viewModel);
 
   @override
   Future<void> refreshAnimation() async {
-    widget.animListener(
-      MvvmContext(context, animationsControllers: _controller),
-      presenter,
-      presenter.viewModel,
-    );
+    animListener(mvvmContext, presenter, presenter.viewModel);
   }
 }
-
-//class MultipleAnimatedMvvmContent<P extends Presenter, M extends MVVMModel> extends MVVMContent {
-//
-//  final MvvmAnimationsControllerBuilder multipleAnimController;
-//
-//  MultipleAnimatedMvvmContent({
-//    Key key,
-//    @required MvvmContentBuilder<P, M> builder,
-//    @required this.multipleAnimController,
-//  }) : super(key: key, builder: builder);
-//
-//  @override
-//  _MVVMMultipleTickerProviderContentState<P, M> createState() =>
-//    _MVVMMultipleTickerProviderContentState<P, M>(_builder, this.multipleAnimController);
-//}
-//
-//
-//class _MVVMMultipleTickerProviderContentState<P extends Presenter, M extends MVVMModel> extends _MVVMContentState with TickerProviderStateMixin {
-//
-//  final MvvmAnimationsControllerBuilder animationControllerBuilder;
-//  List<AnimationController> _controllers;
-//
-//  _MVVMMultipleTickerProviderContentState(MvvmContentBuilder<P, M> builder, this.animationControllerBuilder) : super(builder);
-//
-//  @override
-//  void didChangeDependencies() {
-//    super.didChangeDependencies();
-//    _controllers = animationControllerBuilder(this);
-//  }
-//
-//  @override
-//  Widget build(BuildContext context) =>
-//    _builder(MvvmContext(context, animationsControllers: _controllers), _presenter, this._presenter.viewModel);
-//
-//  @override
-//  refreshAnimation() async {
-//    // TODO: implement refreshAnimation
-//    throw UnimplementedError();
-//  }
-//
-//
-//}